refactor: separate out resize indicator
